+++ engine/docs/changelog.md
@@ -1,4 +1,8 @@
 # TBD
+### Breaking Changes
+* Remove all the following fields from the `EnclaveInfo` object returned by the API:
+    * `network_id`
+    * `network_cidr`
 
 # 1.18.1
 ### Features